Update Maven and Java versions in Dockerfile and mvnw files
The Dockerfile and mvnw files were updated to use newer versions of Maven and Java. In the Dockerfile, maven:3.8-eclipse-temurin-19 was changed to maven:3.9-eclipse-temurin-21, and eclipse-temurin:19-jdk to eclipse-temurin:21-jdk. Similarly, in the mvnw files, the Maven version was updated from 3.2.0 to 3.9.6, and the Java versions were updated correspondingly. The Docker build command was also updated to include the --platform linux/arm64 option.